Added the Tour Class and refactoed the code so the view does not receive HTTP requests
