+++ createTour/urls.py
@@ -1,10 +1,10 @@
 from django.conf.urls import url
-
 from . import views
+from createTour import controllers
 
 urlpatterns = [
-    url(r'^$', views.details, name='selected details'),
-    url(r'^itinerary_update_delete', views.itineraryUpdateDeleteView),
+    url(r'^$', controllers.tourController, name='selected details'),
+    url(r'^itinerary_update_delete', controllers.itineraryUpdateDelete),
     url(r'^hotels/', views.hotels, name="hotels"),
 
 ]
